microsphere-dynamic-jdbc-spring-boot-starter/src/main/java/io/microsphere/dynamic/jdbc/spring/boot/config/AbstractConfigPostProcessor.java
package io.microsphere.dynamic.jdbc.spring.boot.config;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.BeansException;
import org.springframework.beans.factory.BeanClassLoaderAware;
import org.springframework.context.ApplicationContext;
import org.springframework.context.ApplicationContextAware;
import org.springframework.context.ConfigurableApplicationContext;
import org.springframework.context.EnvironmentAware;
import org.springframework.core.env.Environment;

/**
 * Abstract {@link ConfigPostProcessor}
 * 
 * @author <a href="mailto:mercyblitz@gmail.com">Mercy<a/>
 * @since 1.0.0
 */
public abstract class AbstractConfigPostProcessor implements ConfigPostProcessor, ApplicationContextAware, EnvironmentAware, BeanClassLoaderAware {

    protected final Logger logger = LoggerFactory.getLogger(getClass());

    protected Environment environment;

    protected ClassLoader classLoader;

    protected ConfigurableApplicationContext context;

    private int order = 0;

    @Override
    public int getOrder() {
        return this.order;
    }

    public void setOrder(int order) {
        this.order = order;
    }

    @Override
    public void setApplicationContext(ApplicationContext context) throws BeansException {
        this.context = (ConfigurableApplicationContext) context;
    }

    @Override
    public void setEnvironment(Environment environment) {
        this.environment = environment;
    }

    @Override
    public void setBeanClassLoader(ClassLoader classLoader) {
        this.classLoader = classLoader;
    }
}
